add changes based on code review
