+++ packages/frontend/src/components/ErrorResult/GenericErrorResult.tsx
@@ -1,11 +1,10 @@
 import { IJSONObject } from '@plumber/types'
 
-import { useCallback, useState } from 'react'
-import { Box, Collapse, Text } from '@chakra-ui/react'
-import { Button, Infobox } from '@opengovsg/design-system-react'
+import { Box, Text } from '@chakra-ui/react'
+import { Infobox } from '@opengovsg/design-system-react'
 
-import JSONViewer from '@/components/JSONViewer'
-import { SUPPORT_FORM_LINK } from '@/config/urls'
+import ErrorDetailsCollapse from './ErrorDetailsCollapse'
+import SupportContactMessage from './SupportContactMessage'
 
 interface GenericErrorResultProps {
   errorDetails: IJSONObject
@@ -14,10 +13,6 @@ interface GenericErrorResultProps {
 
 export default function GenericErrorResult(props: GenericErrorResultProps) {
   const { errorDetails, isTestRun } = props
-  const [isOpen, setIsOpen] = useState(false)
-  const toggleDropdown = useCallback(() => {
-    setIsOpen((value) => !value)
-  }, [])
 
   return (
     <Infobox variant="error" borderRadius="lg">
@@ -27,39 +22,12 @@ export default function GenericErrorResult(props: GenericErrorResultProps) {
         </Text>
 
         <Text textStyle="body-1">
-          <Text textStyle="body-1">
-            {`Check if you have configured ${
-              isTestRun ? 'the steps above' : 'this step'
-            } correctly and retest.`}
-          </Text>
-          <Box
-            marginTop={4}
-            borderTop="1px solid #E0E0E0"
-            fontSize="0.8rem"
-            opacity={0.8}
-            w="full"
-          >
-            If this error still persists, contact{' '}
-            <a href={SUPPORT_FORM_LINK} target="_blank" rel="noreferrer">
-              Plumber support
-            </a>
-            .
-          </Box>
-          <Button
-            onClick={toggleDropdown}
-            variant="link"
-            size="sm"
-            sx={{ textDecoration: 'underline' }}
-            mt={2}
-          >
-            View error details below.
-          </Button>
+          {`Check if you have configured ${
+            isTestRun ? 'the steps above' : 'this step'
+          } correctly and retest.`}
 
-          <Box>
-            <Collapse in={isOpen}>
-              <JSONViewer data={errorDetails}></JSONViewer>
-            </Collapse>
-          </Box>
+          <SupportContactMessage />
+          <ErrorDetailsCollapse details={errorDetails} />
         </Text>
       </Box>
     </Infobox>
